exploring a new direction for the style guide and linting enforcement
what if we remove dependency on Prettier, and solely rely on ESLint?

it seems we will lose some strictness in our enforcement, but possibly
gain a better development experience in this codebase. Prettier reprints
the entire program following its own style (subject to a few config options)
but its style is often subpar for our needs. it also can be tricky
for IDEs, extensions, and CI to integrate with. we've had our fair share
of issues and exploring this direction will allow us to understand what
we might gain and whether its worth it.